Refactored gae-js-migrations codebase to be neater, and created __test folder to hold test utilities
