+++ packages/gae-js-firestore/src/firestore/firestore-loader.test.ts
@@ -1,6 +1,6 @@
 import { FirestoreLoader } from "./firestore-loader";
 import { Firestore } from "@google-cloud/firestore";
-import { connectFirestore, deleteCollection } from "./test-utils";
+import { connectFirestore, deleteCollection } from "../__test/test-utils";
 import assert from "assert";
 
 describe("FirestoreLoader", () => {
