+++ packages/gae-js-migrations/src/index.ts
@@ -1,5 +1,3 @@
 export * from "./auto-migration";
-export * from "./migration.bootstrapper";
-export * from "./mutex.service";
-export * from "./mutex-unavailable-error";
-export * from "./mutexes.repository";
+export * from "./migration";
+export * from "./mutex";
