+++ packages/gae-js-migrations/src/migration.test.ts
@@ -1,10 +1,9 @@
-import { runMigrations } from "./migration.bootstrapper";
+import { runMigrations } from "./migration";
 import { AutoMigration } from "./auto-migration";
 import { migrationResultsRepository } from "./migration-results.repository";
 import { newTimestampedEntity } from "@mondomob/gae-js-firestore";
-import { transactional, useFirestoreTest } from "./test-utils";
-import { mutexServiceProvider } from "./mutex.service";
-import { mutexesRepository } from "./mutexes.repository";
+import { transactional, useFirestoreTest } from "./__test/test-utils";
+import { mutexServiceProvider, mutexesRepository } from "./mutex";
 
 describe("runMigrations", () => {
   useFirestoreTest(["mutexes", "migrations"]);
