+++ packages/gae-js-migrations/tsconfig.json
@@ -6,6 +6,7 @@
   },
   "include": ["src/**/*"],
   "references": [
-    { "path": "../gae-js-core/tsconfig.prod.json" }
+    { "path": "../gae-js-core/tsconfig.prod.json" },
+    { "path": "../gae-js-firestore/tsconfig.prod.json" }
   ]
 }
